Refactor group name validation
