+++ backend/src/core/entity/Group.ts
@@ -1,3 +1,5 @@
+import type { Result } from "src/utils";
+
 export class Group {
   /** Group's id */
   public readonly id: number;
@@ -21,5 +23,17 @@ export class Group {
     this.name = name;
     this.membersId = membersId;
     this.transactionsId = transactionsId;
+  }
+
+  static validateName(name: any): Result<string> {
+    if (typeof name !== "string") {
+      return { success: false, error: "Group name should be a string" };
+    }
+
+    if (name.length === 0) {
+      return { success: false, error: "Group name should not be empty" };
+    }
+
+    return { success: true, payload: name };
   }
 }
