+++ backend/src/core/usecase/CreateGroup.ts
@@ -1,16 +1,17 @@
-import type { Group } from "../entity/Group";
+import { type Result } from "src/utils";
+import { Group } from "../entity/Group";
 import type { GroupRepository } from "../repository/GroupRepository";
-import type { Result } from "./utils";
 
 export class CreateGroup {
   constructor(private readonly groupRepo: GroupRepository) {}
 
   async execute(name: string): Promise<Result<Group>> {
-    if (name.length === 0) {
-      return { success: false, error: "Group name should not be empty" };
+    const validationResult = Group.validateName(name);
+    if (!validationResult.success) {
+      return validationResult;
     }
 
-    const group = await this.groupRepo.create(name);
+    const group = await this.groupRepo.create(validationResult.payload);
 
     return {
       success: true,
